+++ js/data.js
@@ -1,0 +1,40 @@
+//модуль, который создаёт данные
+const TITLING_TYPES = [
+  'Приветствие 1',
+  'Приветствие 2',
+  'Приветствие 3',
+  'Приветствие 4',
+];
+const HOUSING_TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+];
+const TIMES = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const FEATURING_TYPES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const DESCRIPTION_TYPES =[
+  'Просторная комната',
+  'Вид из окон на море',
+  'Раздельный санузел',
+  'Большая кухня-гостинная',
+  'Удобное расположение от ресепшена',
+];
+const PHOTO_TYPES = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+const PHOTO_AMOUNT = 10;//длина
+export {TITLING_TYPES, HOUSING_TYPES, TIMES, FEATURING_TYPES, DESCRIPTION_TYPES, PHOTO_TYPES, PHOTO_AMOUNT};
